Move collection to dedicated file
